more work on correlations between datasets
